backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
langchain==0.1.16
langchain-openai==0.1.3
supabase==2.4.5
python-multipart==0.0.9
pydantic==2.7.0
python-dotenv==1.0.1
